Refactoring folder structure
