+++ Chess.Engine/Boards.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Chess.EngineCore.Objects;
 
 namespace Chess.EngineCore
 {
