+++ Chess.Engine/Objects/Square.cs
@@ -4,12 +4,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Chess.EngineCore
+namespace Chess.EngineCore.Objects
 {
     public struct Square
     {
-        public Square(){}
-        public int SquareNumber = 0; 
+        public Square() { }
+        public int SquareNumber = 0;
         public string? SquareName = string.Empty;
         public int Value = 0;
     }
