+++ Chess.Ui/Program.cs
@@ -1,8 +1,9 @@
 ﻿
 using Chess.EngineCore;
+using Chess.EngineCore.Objects;
 using System.Text;
 using static Chess.EngineCore.Boards;
-using static Chess.EngineCore.FenParser;
+using static Chess.EngineCore.Helpers.FenParser;
 
 
 var board = new Chess.EngineCore.Boards.Board();
@@ -12,8 +13,8 @@ var board = new Chess.EngineCore.Boards.Board();
 Piece piece = new Piece();
 
 Engine engine = new Engine(board, piece);
-var fenParser = new Chess.EngineCore.FenParser(board, engine, "");
-var fenBuilder = new Chess.EngineCore.FenBuilder(board, engine);
+var fenParser = new Chess.EngineCore.Helpers.FenParser(board, engine, "");
+var fenBuilder = new Chess.EngineCore.Helpers.FenBuilder(board, engine);
 
 board.BuildChessBoard(64);
